35 commits the branch carried that main did not (21 of its 56 were
already in main by patch-id). Nineteen files conflicted; the ones worth
recording are the four where both sides were right about something and
the merge had to pick, plus two where taking either side wholesale would
have shipped a defect.

- finish()'s buffer resets. The branch DELETED them, on the grounds that
  the except clause repeats them and "nothing could observe" the
  duplicates. That premise is false: _report_error notifies the error
  hook, and the hook runs between the assignment and the except clause
  that repeats it. main's TestWhichLayerReportsATruncatedFrame asserts
  exactly that difference, and it fails against the branch's version.
  It also decides what survives a hook that itself raises -- that
  escapes `except ControlFrameProtocolError`, and without the resets the
  decoder keeps the rejected frame's bytes. Kept main's, with the reason
  written down so the next reader does not delete them again. Both
  approaches reach killed==100 on this file; only one of them is also
  correct when the hook misbehaves.

- mcpauth_golden.json. main regenerated the fixture for MCP 2.0
  semantics; the branch carried the pre-regeneration C# mirror, so the
  merged tree held two disagreeing copies and SharedCorpusParityTests
  failed. Re-running gen_mcpauth_golden.py reproduced the TS copy
  byte-identically, which settles which one was stale rather than
  guessing. C# is back to 1651/1651.

- The fan-out timeout assertions. main's copy of the test file carries
  the `>= BACKENDS` fix (SERVER_IMPL narrows BACKENDS on a live-matrix
  leg, and equality failed every one of them), while the branch's
  carries assertions for its own richer timeout message. Taking either
  file wholesale loses one of the two; both are in.

- ci.yml checkout. Took the branch's `fetch-depth: 0`, which its
  protocol-drift step needs, and dropped the `submodules: true` beside
  it -- main dropped that submodule for the published package on
  2026-08-20.

- version-consistency.test.ts. Took the BRANCH's, which replaces the
  literal changelog date with a version-anchored regex. Its reasoning
  ("an assertion edited to make it pass is not one that checks
  anything") is right, and the pin it removes had already cost a CI
  round on main earlier the same day.

- mutation_equivalents.toml. Kept main's. mutmut ids are function-scoped
  and position-dependent, and main's canonical-header change renumbered
  this file's mutants; the branch's entries name the old numbering.

Verified on the merged tree: quality gate clean, all 15 Python legs,
Go, and C# 1651/1651.

+++ packages/provide-uterm-csharp/src/Provide.Uterm/ControlChannel/Codec.cs
@@ -268,7 +268,11 @@ public static class ControlChannelCodec
                 return false;
             }
 
-            value = (value << 4) | d;
+            // (uint), not (long): d is a nibble in 0..15, but a signed int
+            // operand of a long OR is sign-extended, which is CS0675. The
+            // unsigned cast makes the zero-extension explicit and cannot
+            // change the value for a non-negative d.
+            value = (value << 4) | (uint)d;
         }
 
         return true;
